Add file move and rename prediction, other fixes
If a txt file is renamed or moved to another folder outside of the app (not both), the app will detect it and prompt the user to verify changes. Tags will be migrated from the old Textfile instance to the new one if the user indicates that the "new" file is simply the old one moved/renamed. This is done whenever the home page is requested and changes to the files are found.
There will be false positives (e.g. one file deleted and one file created in the same directory): that's why changes are to be verified.
Verification is done with a popup. Popup has one page per possible tag transfer. User can choose to transfer tags or ignore; either way, the old Textfile instance will be deleted.
The purpose of this feature is to make it so that tags are not always lost if files renamed/moved outside of the app (e.g. in the OS filesystem); it's a safeguard measure.

Also fixed a div nesting problem in User's new view, and renamed some variables.

+++ app/models/poros/updater.rb
@@ -1,28 +1,44 @@
+# Refreshes stuff
+
 class Updater
 
   def everything
-    self.home
+    changes = self.home
     self.tags
+    return changes
   end
 
   def home
+    # Setup
+    create_or_update = ModelInstanceUpdater.new
+    textfiles_in_directory = []
+    textfiles_possibly_modified = []
+    textfiles_lost = []
+
     # Generate all model instances from local files
     directory = User.first.home
     paths = Scanner.new.all_file_paths(directory, "txt")
-    create_or_update = ModelInstanceUpdater.new
-    textfiles_in_directory = []
-
     paths.each do |path|
-      textfile = create_or_update.textfile_from_file path
+      data = create_or_update.textfile_from_file path
+      textfile = data[0]
+      new = data[1]
       textfiles_in_directory << textfile
-    end
-
-    # Remove old Textfiles that are not longer in user's designated directory
-    Textfile.all.each do |tf|
-      unless textfiles_in_directory.include? tf
-        tf.destroy
+      if new
+        textfiles_possibly_modified << textfile
       end
     end
+
+    # Track old Textfiles that are not longer in user's designated directory
+    Textfile.all.each do |tf|
+      unless textfiles_in_directory.include? tf
+        textfiles_lost << tf
+      end
+    end
+    potentially_changed = self.predict_modifications(textfiles_possibly_modified, textfiles_lost)
+    textfiles_lost.each do |tf|
+      tf.destroy unless potentially_changed.any? {|hash| hash[:old_file] == tf}
+    end
+    return potentially_changed
   end
 
   def tags
@@ -31,6 +47,35 @@ class Updater
         tag.destroy
       end
     end
+  end
+
+  def predict_modifications(new, old)
+    file = StringConstructor.new
+    potentially_renamed = []
+    potentially_moved = []
+
+    old.each do |old_tf|
+      folder = file.parent_directory(old_tf.location)
+      # Predict file movements (higher priority)
+      if new.any? { |new_tf|
+        new_tf.name == old_tf.name }
+      then
+        potentially_moved << {old_file: old_tf}
+        potentially_moved[-1][:type] = "moved"
+        potentially_moved[-1][:new_files] = new.select {|new_tf| new_tf.name == old_tf.name}
+      # Predict file renames
+      elsif new.any? { |new_tf|
+        folder == file.parent_directory(new_tf.location)}
+      then
+        # Create hashes with old file
+        # => and associated new files that might be the old file but modified
+        potentially_renamed << {old_file: old_tf}
+        potentially_renamed[-1][:type] = "renamed"
+        potentially_renamed[-1][:new_files] = new.select {|tf| file.parent_directory(tf.location) == folder}
+      end
+    end
+
+    return potentially_renamed + potentially_moved
   end
 
 end
